requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
pydantic-settings
PyJWT
openai==0.28.1
